Move switch campus functionality to its own class in response to review.
