+++ app/src/main/java/com/example/campusguide/SwitchCampus.kt
@@ -1,0 +1,30 @@
+package com.example.campusguide
+
+import android.widget.ToggleButton
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.model.LatLng
+
+class SwitchCampus {
+
+    private lateinit var mMap : GoogleMap;
+
+    constructor(toggleBtn: ToggleButton, mMap: GoogleMap) {
+        this.mMap = mMap
+        toggleBtn.setOnCheckedChangeListener{ _, isAtDowntownCampus ->
+            switchCampus(isAtDowntownCampus)
+        }
+    }
+
+    private fun switchCampus(isAtDowntownCampus: Boolean) {
+        if(isAtDowntownCampus){
+            val loyCoord = LatLng(45.458153, -73.640490)
+            mMap.animateCamera(CameraUpdateFactory.newLatLngZoom(loyCoord, 17.0f))
+
+        } else {
+            val sgwCoord = LatLng(45.495792, -73.578096)
+            mMap.animateCamera(CameraUpdateFactory.newLatLngZoom(sgwCoord, 17.0f))
+        }
+    }
+
+}
